+++ src/com/koxudaxi/pydantic/Pydantic.kt
@@ -14,8 +14,8 @@ import com.jetbrains.extensions.QNameResolveContext
 import com.jetbrains.extensions.resolveToElement
 import com.jetbrains.python.PyNames
 import com.jetbrains.python.codeInsight.typing.PyTypingTypeProvider
+import com.jetbrains.python.packaging.PyPackageManagers
 import com.jetbrains.python.psi.*
-import com.jetbrains.python.psi.impl.PyStarArgumentImpl
 import com.jetbrains.python.psi.impl.PyTargetExpressionImpl
 import com.jetbrains.python.psi.resolve.PyResolveContext
 import com.jetbrains.python.psi.resolve.PyResolveUtil
@@ -33,6 +33,11 @@ const val VALIDATOR_Q_NAME = "pydantic.class_validators.validator"
 const val VALIDATOR_SHORT_Q_NAME = "pydantic.validator"
 const val ROOT_VALIDATOR_Q_NAME = "pydantic.class_validators.root_validator"
 const val ROOT_VALIDATOR_SHORT_Q_NAME = "pydantic.root_validator"
+const val FIELD_VALIDATOR_Q_NAME = "pydantic.field_validator"
+const val FIELD_VALIDATOR_SHORT_Q_NAME = "pydantic.functional_validators.field_validator"
+const val MODEL_VALIDATOR_Q_NAME = "pydantic.model_validator"
+const val MODEL_VALIDATOR_SHORT_Q_NAME = "pydantic.functional_validators.model_validator"
+
 const val SCHEMA_Q_NAME = "pydantic.schema.Schema"
 const val FIELD_Q_NAME = "pydantic.fields.Field"
 const val DATACLASS_FIELD_Q_NAME = "dataclasses.field"
@@ -41,6 +46,9 @@ const val DEPRECATED_SCHEMA_Q_NAME = "pydantic.fields.Schema"
 const val BASE_SETTINGS_Q_NAME = "pydantic.env_settings.BaseSettings"
 const val VERSION_Q_NAME = "pydantic.version.VERSION"
 const val BASE_CONFIG_Q_NAME = "pydantic.main.BaseConfig"
+const val CONFIG_DICT_Q_NAME = "pydantic.config.ConfigDict"
+const val CONFIG_DICT_SHORT_Q_NAME = "pydantic.ConfigDict"
+const val CONFIG_DICT_DEFAULTS_Q_NAME = "pydantic._internal._config.config_defaults"
 const val DATACLASS_MISSING = "dataclasses.MISSING"
 const val CON_BYTES_Q_NAME = "pydantic.types.conbytes"
 const val CON_DECIMAL_Q_NAME = "pydantic.types.condecimal"
@@ -75,6 +83,12 @@ val VERSION_QUALIFIED_NAME = QualifiedName.fromDottedString(VERSION_Q_NAME)
 
 val BASE_CONFIG_QUALIFIED_NAME = QualifiedName.fromDottedString(BASE_CONFIG_Q_NAME)
 
+val CONFIG_DICT_QUALIFIED_NAME = QualifiedName.fromDottedString(CONFIG_DICT_Q_NAME)
+
+val CONFIG_DICT_DEFAULTS_QUALIFIED_NAME = QualifiedName.fromDottedString(CONFIG_DICT_DEFAULTS_Q_NAME)
+
+val CONFIG_DICT_SHORT_QUALIFIED_NAME = QualifiedName.fromDottedString(CONFIG_DICT_SHORT_Q_NAME)
+
 val BASE_MODEL_QUALIFIED_NAME = QualifiedName.fromDottedString(BASE_MODEL_Q_NAME)
 
 val VALIDATOR_QUALIFIED_NAME = QualifiedName.fromDottedString(VALIDATOR_Q_NAME)
@@ -84,6 +98,14 @@ val VALIDATOR_SHORT_QUALIFIED_NAME = QualifiedName.fromDottedString(VALIDATOR_SH
 val ROOT_VALIDATOR_QUALIFIED_NAME = QualifiedName.fromDottedString(ROOT_VALIDATOR_Q_NAME)
 
 val ROOT_VALIDATOR_SHORT_QUALIFIED_NAME = QualifiedName.fromDottedString(ROOT_VALIDATOR_SHORT_Q_NAME)
+
+val FIELD_VALIDATOR_QUALIFIED_NAME = QualifiedName.fromDottedString(FIELD_VALIDATOR_Q_NAME)
+
+val FIELD_VALIDATOR_SHORT_QUALIFIED_NAME = QualifiedName.fromDottedString(FIELD_VALIDATOR_SHORT_Q_NAME)
+
+val MODEL_VALIDATOR_QUALIFIED_NAME = QualifiedName.fromDottedString(MODEL_VALIDATOR_Q_NAME)
+
+val MODEL_VALIDATOR_SHORT_QUALIFIED_NAME = QualifiedName.fromDottedString(MODEL_VALIDATOR_SHORT_Q_NAME)
 
 val DATA_CLASS_QUALIFIED_NAME = QualifiedName.fromDottedString(DATA_CLASS_Q_NAME)
 
@@ -101,6 +123,17 @@ val VALIDATOR_QUALIFIED_NAMES = listOf(
     VALIDATOR_SHORT_QUALIFIED_NAME,
     ROOT_VALIDATOR_QUALIFIED_NAME,
     ROOT_VALIDATOR_SHORT_QUALIFIED_NAME
+)
+
+val V2_VALIDATOR_QUALIFIED_NAMES = listOf(
+    VALIDATOR_QUALIFIED_NAME,
+    VALIDATOR_SHORT_QUALIFIED_NAME,
+    ROOT_VALIDATOR_QUALIFIED_NAME,
+    ROOT_VALIDATOR_SHORT_QUALIFIED_NAME,
+    FIELD_VALIDATOR_QUALIFIED_NAME,
+    FIELD_VALIDATOR_SHORT_QUALIFIED_NAME,
+    MODEL_VALIDATOR_QUALIFIED_NAME,
+    MODEL_VALIDATOR_SHORT_QUALIFIED_NAME
 )
 
 val VERSION_SPLIT_PATTERN: Pattern = Pattern.compile("[.a-zA-Z]")!!
@@ -131,10 +164,16 @@ val CONFIG_TYPES = mapOf(
     "allow_mutation" to ConfigType.BOOLEAN,
     "frozen" to ConfigType.BOOLEAN,
     "keep_untouched" to ConfigType.LIST_PYTYPE,
-    "extra" to ConfigType.EXTRA
+    "extra" to ConfigType.EXTRA,
+    "populate_by_name" to ConfigType.BOOLEAN,
+    "from_attributes" to ConfigType.BOOLEAN,
 )
 
 const val CUSTOM_ROOT_FIELD = "__root__"
+
+const val MODEL_FIELD_PREFIX = "model_"
+
+const val MODEL_CONFIG_FIELD = "model_config"
 
 fun PyTypedElement.getType(context: TypeEvalContext): PyType? = context.getType(this)
 
@@ -210,7 +249,9 @@ internal fun isDataclassMissing(pyTargetExpression: PyTargetExpression): Boolean
     return pyTargetExpression.qualifiedName == DATACLASS_MISSING
 }
 
-internal val PyFunction.isValidatorMethod: Boolean get() = hasDecorator(this, VALIDATOR_QUALIFIED_NAMES)
+internal fun PyFunction.isValidatorMethod(pydanticVersion: KotlinVersion?): Boolean =
+    hasDecorator(this, if(pydanticVersion.isV2) V2_VALIDATOR_QUALIFIED_NAMES else VALIDATOR_QUALIFIED_NAMES)
+
 
 
 internal val PyClass.isConfigClass: Boolean get() = name == "Config"
@@ -343,16 +384,15 @@ fun getPsiElementByQualifiedName(
     return qualifiedName.resolveToElement(QNameResolveContext(contextAnchor, pythonSdk, context))
 }
 
-fun isValidField(field: PyTargetExpression, context: TypeEvalContext): Boolean {
-    if (field.name?.isValidFieldName != true) return false
+fun isValidField(field: PyTargetExpression, context: TypeEvalContext, isV2: Boolean): Boolean {
+    if (field.name?.isValidFieldName(isV2) != true) return false
 
     val annotationValue = field.annotation?.value ?: return true
     // TODO Support a variable.
     return getQualifiedName(annotationValue, context) != CLASSVAR_Q_NAME
 }
 
-val String.isValidFieldName: Boolean get() = !startsWith('_') || this == CUSTOM_ROOT_FIELD
-
+fun String.isValidFieldName(isV2: Boolean): Boolean = (!startsWith('_') || this == CUSTOM_ROOT_FIELD) && !(isV2 && this.startsWith(MODEL_FIELD_PREFIX))
 
 fun getConfigValue(name: String, value: Any?, context: TypeEvalContext): Any? {
     if (value is PyReferenceExpression) {
@@ -399,6 +439,7 @@ fun validateConfig(pyClass: PyClass, context: TypeEvalContext): List<PsiElement>
     return results
 }
 
+
 fun getConfig(
     pyClass: PyClass,
     context: TypeEvalContext,
@@ -406,7 +447,7 @@ fun getConfig(
     pydanticVersion: KotlinVersion? = null,
 ): HashMap<String, Any?> {
     val config = hashMapOf<String, Any?>()
-    val version = pydanticVersion ?: PydanticCacheService.getVersion(pyClass.project, context)
+    val version = pydanticVersion ?: PydanticCacheService.getVersion(pyClass.project)
     getAncestorPydanticModels(pyClass, false, context)
         .reversed()
         .map { getConfig(it, context, false, version) }
@@ -417,6 +458,31 @@ fun getConfig(
                 }
             }
         }
+    if (version?.isV2 == true) {
+        val configDict = pyClass.findClassAttribute(MODEL_CONFIG_FIELD, false, context)?.findAssignedValue().let {
+            when (it) {
+                is PyReferenceExpression -> {
+                    val targetExpression = getResolvedPsiElements(it, context).firstOrNull() ?: return@let null
+                    (targetExpression as? PyTargetExpression)?.findAssignedValue() ?: return@let null
+                }
+                else -> it
+            }
+        }
+        when (configDict) {
+            is PyDictLiteralExpression -> configDict.elements.forEach { element ->
+                    element.key.text.drop(1).dropLast(1).let { name ->
+                        config[name] = getConfigValue(name, element.value, context)
+                    }
+                }
+            is PyCallExpression -> configDict.arguments.forEach { argument ->
+                    argument.name?.let {name ->
+                        configDict.getKeywordArgument(name)?.let { value ->
+                            config[name] = getConfigValue(name, value, context)
+                        }
+                    }
+                }
+        }
+    }
     pyClass.nestedClasses.firstOrNull { it.isConfigClass }?.let {
         it.classAttributes.forEach { attribute ->
             attribute.findAssignedValue()?.let { value ->
@@ -436,10 +502,18 @@ fun getConfig(
     }
 
     if (setDefault) {
-        DEFAULT_CONFIG.forEach { (key, value) ->
-            if (!config.containsKey(key)) {
-                config[key] = getConfigValue(key, value, context)
+        if (version?.isV2 == true) {
+            PydanticCacheService.getConfigDictDefaults(pyClass.project, context)
+                ?.filterNot { config.containsKey(it.key) }
+                ?.forEach { (name, value) ->
+                    config[name] = value
+                }
             }
+        } else {
+            DEFAULT_CONFIG.forEach { (key, value) ->
+                if (!config.containsKey(key)) {
+                    config[key] = getConfigValue(key, value, context)
+                }
         }
     }
     return config
@@ -469,17 +543,29 @@ fun getPydanticBaseConfig(project: Project, context: TypeEvalContext): PyClass? 
     return getPyClassFromQualifiedName(BASE_CONFIG_QUALIFIED_NAME, project, context)
 }
 
+fun getPydanticConfigDictDefaults(project: Project, context: TypeEvalContext): PyCallExpression? {
+    val targetExpression = getPyTargetExpressionFromQualifiedName(CONFIG_DICT_DEFAULTS_QUALIFIED_NAME, project, context) ?: return null
+    return targetExpression.findAssignedValue() as? PyCallExpression
+}
+
 fun getPydanticBaseModel(project: Project, context: TypeEvalContext): PyClass? {
     return getPyClassFromQualifiedName(BASE_MODEL_QUALIFIED_NAME, project, context)
 }
 
-fun getPyClassFromQualifiedName(qualifiedName: QualifiedName, project: Project, context: TypeEvalContext): PyClass? {
+fun getPsiElementFromQualifiedName(qualifiedName: QualifiedName, project: Project, context: TypeEvalContext): PsiElement? {
     val module = project.modules.firstOrNull() ?: return null
     val pythonSdk = module.pythonSdk
     val contextAnchor = ModuleBasedContextAnchor(module)
-    return qualifiedName.resolveToElement(QNameResolveContext(contextAnchor, pythonSdk, context)) as? PyClass
+    return qualifiedName.resolveToElement(QNameResolveContext(contextAnchor, pythonSdk, context))
 }
 
+fun getPyClassFromQualifiedName(qualifiedName: QualifiedName, project: Project, context: TypeEvalContext): PyClass? {
+    return getPsiElementFromQualifiedName(qualifiedName, project, context) as? PyClass
+}
+
+fun getPyTargetExpressionFromQualifiedName(qualifiedName: QualifiedName, project: Project, context: TypeEvalContext): PyTargetExpression? {
+    return getPsiElementFromQualifiedName(qualifiedName, project, context) as? PyTargetExpression
+}
 fun getPyClassByAttribute(pyPsiElement: PsiElement?): PyClass? {
     return pyPsiElement?.parent?.parent as? PyClass
 }
@@ -662,3 +748,7 @@ fun PyCallableType.getPydanticModel(includeDataclass: Boolean, context: TypeEval
 
 val KotlinVersion?.isV2: Boolean
     get() = this?.isAtLeast(2, 0) == true
+
+val Sdk.pydanticVersion: String?
+    get() = PyPackageManagers.getInstance()
+        .forSdk(this).packages?.find { it.name == "pydantic" }?.version