+++ src/com/koxudaxi/pydantic/PydanticPackageManagerListener.kt
@@ -7,20 +7,18 @@ import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.projectRoots.Sdk
 import com.intellij.openapi.util.Disposer
 import com.jetbrains.python.packaging.PyPackageManager
-import com.jetbrains.python.packaging.PyPackageManagers
 import com.jetbrains.python.sdk.PythonSdkUtil
 import com.jetbrains.python.statistics.sdks
 
 class PydanticPackageManagerListener : PyPackageManager.Listener {
     private fun updateVersion(sdk: Sdk) {
-        val version = PyPackageManagers.getInstance()
-            .forSdk(sdk).packages?.find { it.name == "pydantic" }?.version
+        val version = sdk.pydanticVersion
         ProjectManager.getInstance().openProjects
             .filter { it.sdks.contains(sdk) }
             .forEach {
-                when (version) {
-                    is String -> PydanticCacheService.setVersion(it, version)
-                    else -> PydanticCacheService.clear(it)
+                PydanticCacheService.clear(it)
+                if (version is String) {
+                    PydanticCacheService.getOrPutVersionFromVersionCache(it, version)
                 }
             }
     }