+++ testData/mock/pydanticv2/deprecated/__init__.py
@@ -1,0 +1,1 @@
+from class_validators import validator, root_validator