+++ testSrc/com/koxudaxi/pydantic/PydanticInspectionV2Test.kt
@@ -14,4 +14,19 @@ open class PydanticInspectionV2Test : PydanticInspectionBase(version = "v2") {
         pydanticConfigService.mypyWarnUntypedFields = false
         doTest()
     }
+    fun testValidatorSelf() {
+        doTest()
+    }
+
+    fun testReadOnlyPropertyFrozenConfig() {
+        doTest()
+    }
+
+    fun testReadOnlyPropertyFrozenConfigDict() {
+        doTest()
+    }
+
+    fun testValidators() {
+        doTest()
+    }
 }