+++ testSrc/com/koxudaxi/pydantic/PydanticTestCase.kt
@@ -21,7 +21,7 @@ import com.jetbrains.python.psi.impl.PythonLanguageLevelPusher
 import com.jetbrains.python.sdk.PythonSdkUtil
 import kotlin.test.Test
 
-abstract class PydanticTestCase(version: String = "v1") : UsefulTestCase() {
+abstract class PydanticTestCase(val version: String = "v1") : UsefulTestCase() {
 
     protected var myFixture: CodeInsightTestFixture? = null
 
@@ -82,10 +82,15 @@ abstract class PydanticTestCase(version: String = "v1") : UsefulTestCase() {
                 .also { sdk.sdkModificator.addRoot(it, OrderRootType.CLASSES) }
             libDir.createChildDirectory(null, PyNames.SITE_PACKAGES)
         }
-
+        val parsedVersion = version.split("v")[1].let {
+            when (it) {
+                "18" -> "1.8"
+                else -> it
+            }
+        }
+        PydanticCacheService.setVersion(myFixture!!.project, parsedVersion)
         setLanguageLevel(defaultPythonLanguageLevel)
         InspectionProfileImpl.INIT_INSPECTIONS = true;
-        PydanticCacheService.clear(myFixture!!.project)
     }